# feat: undefined json keys as defaults when using `?columns` and a `Prefer` header