--- a/src/PostgREST/ApiRequest/Preferences.hs
+++ b/src/PostgREST/ApiRequest/Preferences.hs
@@ -9,7 +9,7 @@
 module PostgREST.ApiRequest.Preferences
   ( Preferences(..)
   , PreferCount(..)
-  , PreferDefaults(..)
+  , PreferUndefinedKeys(..)
   , PreferParameters(..)
   , PreferRepresentation(..)
   , PreferResolution(..)
@@ -34,6 +34,7 @@ import Protolude
 -- >>> deriving instance Show PreferParameters
 -- >>> deriving instance Show PreferCount
 -- >>> deriving instance Show PreferTransaction
+-- >>> deriving instance Show PreferUndefinedKeys
 -- >>> deriving instance Show Preferences
 
 -- | Preferences recognized by the application.
@@ -44,7 +45,7 @@ data Preferences
     , preferParameters     :: Maybe PreferParameters
     , preferCount          :: Maybe PreferCount
     , preferTransaction    :: Maybe PreferTransaction
-    , preferDefaults       :: Maybe PreferDefaults
+    , preferUndefinedKeys  :: Maybe PreferUndefinedKeys
     }
 
 -- |
@@ -59,6 +60,7 @@ data Preferences
 --     , preferParameters = Nothing
 --     , preferCount = Just ExactCount
 --     , preferTransaction = Nothing
+--     , preferUndefinedKeys = Nothing
 --     }
 --
 -- Multiple headers can also be used:
@@ -70,6 +72,7 @@ data Preferences
 --     , preferParameters = Nothing
 --     , preferCount = Just ExactCount
 --     , preferTransaction = Nothing
+--     , preferUndefinedKeys = Nothing
 --     }
 --
 -- If a preference is set more than once, only the first is used:
@@ -94,14 +97,14 @@ data Preferences
 --
 -- Preferences can be separated by arbitrary amounts of space, lower-case header is also recognized:
 --
--- >>> pPrint $ fromHeaders [("prefer", "count=exact,    tx=commit   ,return=representation , defaults=apply")]
+-- >>> pPrint $ fromHeaders [("prefer", "count=exact,    tx=commit   ,return=representation , undefined-keys=apply-defaults")]
 -- Preferences
 --     { preferResolution = Nothing
 --     , preferRepresentation = Full
 --     , preferParameters = Nothing
 --     , preferCount = Just ExactCount
 --     , preferTransaction = Just Commit
---     , preferDefaults = Just ApplyDefaults
+--     , preferUndefinedKeys = Just ApplyDefaults
 --     }
 --
 fromHeaders :: [HTTP.Header] -> Preferences
@@ -112,7 +115,7 @@ fromHeaders headers =
     , preferParameters = parsePrefs [SingleObject, MultipleObjects]
     , preferCount = parsePrefs [ExactCount, PlannedCount, EstimatedCount]
     , preferTransaction = parsePrefs [Commit, Rollback]
-    , preferDefaults = parsePrefs [ApplyDefaults, IgnoreDefaults]
+    , preferUndefinedKeys = parsePrefs [ApplyDefaults, IgnoreDefaults]
     }
   where
     prefHeaders = filter ((==) HTTP.hPrefer . fst) headers
@@ -211,14 +214,14 @@ instance ToAppliedHeader PreferTransaction
 
 -- |
 -- How to handle the insertion/update when the keys specified in ?columns are not present
--- in the json body. 
-data PreferDefaults
+-- in the json body.
+data PreferUndefinedKeys
   = ApplyDefaults  -- ^ Use the default column value for the unspecified keys.
-  | IgnoreDefaults -- ^ Inserts: null values / Updates: the keys are not SET to any value 
+  | IgnoreDefaults -- ^ Inserts: null values / Updates: the keys are not SET to any value
   deriving Eq
 
-instance ToHeaderValue PreferDefaults where
-  toHeaderValue ApplyDefaults  = "defaults=apply"
-  toHeaderValue IgnoreDefaults = "defaults=ignore"
+instance ToHeaderValue PreferUndefinedKeys where
+  toHeaderValue ApplyDefaults  = "undefined-keys=apply-defaults"
+  toHeaderValue IgnoreDefaults = "undefined-keys=ignore-defaults"
 
-instance ToAppliedHeader PreferDefaults
+instance ToAppliedHeader PreferUndefinedKeys