… CHANGELOG.md
… test/pgbench/1567/old.sql
… test/pgbench/README.md
… test/spec/Feature/Query/InsertSpec.hs
… test/spec/Feature/Query/QuerySpec.hs
… test/spec/Feature/Query/UpdateSpec.hs
… test/spec/fixtures/data.sql
… test/spec/fixtures/schema.sql