Closes #1567.

(`field-with_sep` has default 1 on the following)

- Allows bulk insert with defaults

```bash
http POST "localhost:3000/complex_items?columns=id,name,field-with_sep,arr_data" "Prefer: return=representation" <<JSON
[
  {"id": 4, "name": "Vier"},
  {"id": 5, "name": "Funf", "arr_data": null},
  {"id": 6, "name": "Sechs", "field-with_sep": 6, "arr_data": [1,2,3]}
]
JSON

[
    { "arr_data": null, "field-with_sep": 1, "id": 4, "name": "Vier", "settings": null },
    { "arr_data": null, "field-with_sep": 1, "id": 5, "name": "Funf", "settings": null },
    { "arr_data": [ 1, 2, 3 ], "field-with_sep": 6, "id": 6, "name": "Sechs", "settings": null }
]
```

- Allows updating with defaults

```bash
http PATCH "/complex_items?id=eq.3&columns=name,field-with_sep" "Prefer: return=representation" <<JSON
{"name": "Tres"}
JSON


[ { "arr_data": [ 1, 2, 3 ], "field-with_sep": 1, "id": 3, "name": "Tres", "settings": { "foo": { "bar": "baz", "int": 1 } } } ]
```

- It works on views as well assuming they have a default set up using: `ALTER VIEW test.complex_items_view ALTER COLUMN name SET DEFAULT 'Default'`

## Pending

- [x] Optionally apply the feature with a Prefer header(since this new insert query is more expensive)
- [x] CHANGELOG