@steve-chavez 2023-02-19
> It works on views as well assuming they have a default set up using: ALTER VIEW test.complex_items_view ALTER COLUMN name SET DEFAULT 'Default'

The above is not ideal. When omitting a column on a view INSERT:

```sql
INSERT INTO test.complex_items_view(id,name) VALUES (11, 'Eleven') RETURNING *;
 id |  name  | settings | arr_data | field-with_sep
----+--------+----------+----------+----------------
 11 | Eleven |          |          |              1
```

PostgreSQL correctly infers the `field-with_sep` default which is defined on the source table but when leaving an undefined key on the json we don't infer the default. I'm leaving this for a later enhancement.