@steve-chavez 2023-02-20
> Check if it makes sense for upsert(might need additional logic)

So I was thinking that doing this:

(consider that `id:3` already exists)

```bash
http POST "localhost:3000/complex_items?columns=id,name,field-with_sep,arr_data" "Prefer: return=representation, resolution=merge-duplicates" <<JSON
[
  {"id": 3, "name": "Drei"},
  {"id": 4, "name": "Vier"},
  {"id": 5, "name": "Funf", "arr_data": null},
  {"id": 6, "name": "Sechs", "field-with_sep": 6, "arr_data": [1,2,3]}
]
JSON
```

Would only UPDATE the `name` column for the `id: 3` column, currently it updates the `field-with_sep` to the default and `arr_data` to `null`.

I was thinking to do this automatically but:

- doing so would not allow to update a column to a default on upsert
- it'd be better to be explicit about the updated columns: https://github.com/PostgREST/postgrest/issues/2602#issuecomment-1382753901