@steve-chavez 2023-02-21
So I just made a pgbench test on the old insert query compared to the new one:

### old

```bash
$ postgrest-with-postgresql-15 -f test/pgbench/fixtures.sql pgbench -T 30 -f test/pgbench/1567/old.sql

pgbench (15.1)
starting vacuum...pgbench: error: ERROR:  relation "pgbench_branches" does not exist
pgbench: detail: (ignoring this error and continuing anyway)
pgbench: error: ERROR:  relation "pgbench_tellers" does not exist
pgbench: detail: (ignoring this error and continuing anyway)
pgbench: error: ERROR:  relation "pgbench_history" does not exist
pgbench: detail: (ignoring this error and continuing anyway)
end.
transaction type: test/pgbench/1567/old.sql
scaling factor: 1
query mode: simple
number of clients: 1
number of threads: 1
maximum number of tries: 1
duration: 30 s
number of transactions actually processed: 146236
number of failed transactions: 0 (0.000%)
latency average = 0.205 ms
initial connection time = 1.234 ms
tps = 4874.710773 (without initial connection time)
```

### new

```
$ postgrest-with-postgresql-15 -f test/pgbench/fixtures.sql pgbench -T 30 -f test/pgbench/1567/new.sql

pgbench (15.1)
starting vacuum...pgbench: error: ERROR:  relation "pgbench_branches" does not exist
pgbench: detail: (ignoring this error and continuing anyway)
pgbench: error: ERROR:  relation "pgbench_tellers" does not exist
pgbench: detail: (ignoring this error and continuing anyway)
pgbench: error: ERROR:  relation "pgbench_history" does not exist
pgbench: detail: (ignoring this error and continuing anyway)
end.
transaction type: test/pgbench/1567/old.sql
scaling factor: 1
query mode: simple
number of clients: 1
number of threads: 1
maximum number of tries: 1
duration: 30 s
number of transactions actually processed: 146236
number of failed transactions: 0 (0.000%)
latency average = 0.205 ms
initial connection time = 1.234 ms
tps = 4874.710773 (without initial connection time)
pgbench (15.1)
starting vacuum...pgbench: error: ERROR:  relation "pgbench_branches" does not exist
pgbench: detail: (ignoring this error and continuing anyway)
pgbench: error: ERROR:  relation "pgbench_tellers" does not exist
pgbench: detail: (ignoring this error and continuing anyway)
pgbench: error: ERROR:  relation "pgbench_history" does not exist
pgbench: detail: (ignoring this error and continuing anyway)
end.
transaction type: test/pgbench/1567/new.sql
scaling factor: 1
query mode: simple
number of clients: 1
number of threads: 1
maximum number of tries: 1
duration: 30 s
number of transactions actually processed: 121645
number of failed transactions: 0 (0.000%)
latency average = 0.247 ms
initial connection time = 1.253 ms
tps = 4054.990532 (without initial connection time)
```

## results

So there's `16.8%` loss in perf.

>  Shortcut for tables that don't have defaults(since this new insert query is more expensive)

The above is adequate but I'm also thinking maybe we should add a new `Prefer` header for this feature.