@steve-chavez 2023-03-08
> Like Prefer: infer-defaults=true.

The above wouldn't be right if this feature is JSON specific, it would have to be like:

- `Content-Type: application/vnd.pgrst.object+json; apply-defaults=true`
- `Content-Type: application/vnd.pgrst.array+json; apply-defaults=true`

Maybe it could be added to CSV as well though.. in which case the `Prefer` would not be incorrect. It's also less ergonomic to switch the whole content type for an optional behavior.

Also both options could be done.