@steve-chavez 2023-03-09
So I've added a `Prefer: defaults=apply/ignore` header. Maybe to be more precise this should be `Prefer: undefined-keys = apply-defaults/ignore-defaults` .