@steve-chavez 2023-03-09
Having to add both `columns` and the `Prefer` header also seems like a complicated interface. Maybe it won't matter later when we get rid of the Haskell json processing.