@steve-chavez 2023-03-21
So since this feature requires `jsonb`, we no longer would be able to insert json with duplicates keys, if we were to switch completely to `jsonb`. This would be a breaking change.

To avoid this, I'm only switching to `jsonb` whenever the new Prefer header is used. I've also added tests on for this(on https://github.com/PostgREST/postgrest/pull/2672/commits/dc3335798f884f95a1b4e77bf2ff5947d9005b93).