@steve-chavez 2023-03-21
> Having to add both columns and the Prefer header also seems like a complicated interface.

To avoid complicating the interface, I've enabled this Prefer header to work without `?columns`. I see no harm in doing so. We can document that this is only required with `?columns`.