@steve-chavez 2023-03-22
Tried to appease codecov with the method mentioned on https://github.com/PostgREST/postgrest/pull/2671/, it cleared up one warning but not the other 2, no idea how to solve those.

Will leave it like that, they're only record fields which codecov doesn't understand.