@steve-chavez 2023-03-27
> What about something like Prefer: missing-values=apply-defaults|ignore-defaults or Prefer: missing=default|null. I actually like the latter more.

Great idea! It looks more generic.

Doing that on https://github.com/PostgREST/postgrest/pull/2723.